fix: copy hook
